# api-documentation.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "InternshipPortal.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Internship/{id}": {
      "put": {
        "tags": [
          "Internship"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInternshipDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInternshipDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInternshipDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Internship"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Internship"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/StudentProfile": {
      "post": {
        "tags": [
          "StudentProfile"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudentProfileDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudentProfileDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudentProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "StudentProfile"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentProfileDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentProfileDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/StudentProfile/upload-resume": {
      "post": {
        "tags": [
          "StudentProfile"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/StudentProfile/upload-profile-image": {
      "post": {
        "tags": [
          "StudentProfile"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/TrainingMaterial/upload": {
      "post": {
        "tags": [
          "TrainingMaterial"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "InternshipId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "Title": {
                    "type": "string"
                  },
                  "Description": {
                    "type": "string"
                  },
                  "AcceptedStudentsOnly": {
                    "type": "boolean"
                  },
                  "CompletedStudentsOnly": {
                    "type": "boolean"
                  },
                  "File": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "InternshipId": {
                  "style": "form"
                },
                "Title": {
                  "style": "form"
                },
                "Description": {
                  "style": "form"
                },
                "AcceptedStudentsOnly": {
                  "style": "form"
                },
                "CompletedStudentsOnly": {
                  "style": "form"
                },
                "File": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/TrainingMaterial/internship/{internshipId}": {
      "get": {
        "tags": [
          "TrainingMaterial"
        ],
        "parameters": [
          {
            "name": "internshipId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/TrainingMaterial/{materialId}": {
      "delete": {
        "tags": [
          "TrainingMaterial"
        ],
        "parameters": [
          {
            "name": "materialId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateStudentProfileDto": {
        "type": "object",
        "properties": {
          "collegeName": {
            "type": "string",
            "nullable": true
          },
          "department": {
            "type": "string",
            "nullable": true
          },
          "cgpa": {
            "type": "number",
            "format": "double"
          },
          "backlogs": {
            "type": "integer",
            "format": "int32"
          },
          "graduationYear": {
            "type": "integer",
            "format": "int32"
          },
          "skills": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateInternshipDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "companyName": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "stipend": {
            "type": "number",
            "format": "double"
          },
          "durationInMonths": {
            "type": "integer",
            "format": "int32"
          },
          "minimumCGPA": {
            "type": "number",
            "format": "double"
          },
          "allowedBacklogs": {
            "type": "integer",
            "format": "int32"
          },
          "requiredSkills": {
            "type": "string",
            "nullable": true
          },
          "allowedDepartments": {
            "type": "string",
            "nullable": true
          },
          "graduationYear": {
            "type": "integer",
            "format": "int32"
          },
          "totalSeats": {
            "type": "integer",
            "format": "int32"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "registrationDeadline": {
            "type": "string",
            "format": "date-time"
          },
          "coverImageUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateStudentProfileDto": {
        "type": "object",
        "properties": {
          "collegeName": {
            "type": "string",
            "nullable": true
          },
          "department": {
            "type": "string",
            "nullable": true
          },
          "cgpa": {
            "type": "number",
            "format": "double"
          },
          "backlogs": {
            "type": "integer",
            "format": "int32"
          },
          "graduationYear": {
            "type": "integer",
            "format": "int32"
          },
          "skills": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}